requirements.txt
torch
torchvision
transformers
torchaudio
numpy
pandas
matplotlib
scikit-learn
tqdm
einops
Pillow
